move sortable header to permanent location
