+++ src/frontend/src/views/Data/components/SamplesView/components/SamplesTable/index.tsx
@@ -23,7 +23,7 @@ import { IdMap } from "src/common/utils/dataTransforms";
 import { datetimeWithTzToLocalDate } from "src/common/utils/timeUtils";
 import { LineageTooltip } from "./components/LineageTooltip";
 import { DefaultCell } from "./components/DefaultCell";
-import { SortableHeader } from "./components/SortableHeader";
+import { SortableHeader } from "src/views/Data/components/SortableHeader";
 import { StyledCellBasic, StyledPrivateId, StyledTableRow } from "./style";
 import { EmptyTable } from "src/views/Data/components/EmptyState";
