+++ src/frontend/src/views/Data/components/TreesView/components/TreesTable/style.ts
@@ -1,6 +1,5 @@
 import styled from "@emotion/styled";
-// TODO-TR: import
-import { SortableHeader } from "../../../SamplesView/components/SamplesTable/components/SortableHeader";
+import { SortableHeader } from "src/views/Data/components/SortableHeader";
 
 export const StyledSortableHeader = styled(SortableHeader)`
   width: "auto";
